+++ spec/legion/extensions/telemetry/actors/collector_spec.rb
@@ -2,16 +2,6 @@
 
 require 'spec_helper'
 
-unless defined?(Legion::Extensions::Actors::Every)
-  module Legion
-    module Extensions
-      module Actors
-        class Every; end
-      end
-    end
-  end
-end
-
 require 'legion/extensions/telemetry/actors/collector'
 
 RSpec.describe Legion::Extensions::Telemetry::Actor::Collector do