+++ spec/legion/extensions/telemetry/transport/messages/telemetry_message_spec.rb
@@ -2,25 +2,11 @@
 
 require 'spec_helper'
 
-unless defined?(Legion::Transport::Message)
-  module Legion
-    module Transport
-      class Message
-        def routing_key = nil
-      end
-
-      class Exchange
-        def exchange_name = nil
-      end
-    end
-  end
-end
-
 require 'legion/extensions/telemetry/transport/exchanges/sessions'
 require 'legion/extensions/telemetry/transport/messages/telemetry_message'
 
 RSpec.describe Legion::Extensions::Telemetry::Transport::Messages::TelemetryMessage do
-  subject(:message) { described_class.new }
+  subject(:message) { described_class.allocate }
 
   it 'routes to telemetry.sessions.process' do
     expect(message.routing_key).to eq('telemetry.sessions.process')