+++ spec/legion/extensions/telemetry/transport/queues/sessions_process_spec.rb
@@ -2,20 +2,10 @@
 
 require 'spec_helper'
 
-unless defined?(Legion::Transport::Queue)
-  module Legion
-    module Transport
-      class Queue
-        def queue_name = nil
-      end
-    end
-  end
-end
-
 require 'legion/extensions/telemetry/transport/queues/sessions_process'
 
 RSpec.describe Legion::Extensions::Telemetry::Transport::Queues::SessionsProcess do
-  subject(:queue) { described_class.new }
+  subject(:queue) { described_class.allocate }
 
   it 'has queue name telemetry.sessions.process' do
     expect(queue.queue_name).to eq('telemetry.sessions.process')